# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book Management API",
        "description": "A comprehensive API for managing books and book loans",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/loans": {
            "get": {
                "tags": [
                    "Book Loans"
                ],
                "summary": "Get list of book loans",
                "description": "Retrieve a paginated list of book loans with optional filters",
                "operationId": "1d16c7f613519d93c4791cebcab2f55f",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filter loans by user ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BookLoan"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Book Loans"
                ],
                "summary": "Create a new book loan",
                "description": "Create a new book loan for a user",
                "operationId": "adca2cb247c9ae512fe67c9b65222ee5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "book_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "book_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "expected_return_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-01-22T10:30:00.000000Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book loaned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BookLoan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or book not available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/loans/{id}": {
            "get": {
                "tags": [
                    "Book Loans"
                ],
                "summary": "Get book loan by ID",
                "description": "Retrieve a specific book loan by its ID",
                "operationId": "42fc8c455062d8f933479eea82111f88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book loan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book loan found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BookLoan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book loan not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Book Loans"
                ],
                "summary": "Delete book loan",
                "description": "Delete a book loan record",
                "operationId": "ce34f01761df6f1b2f56c7b0d49c217a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book loan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book loan deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book loan deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book loan not found"
                    }
                }
            }
        },
        "/api/loans/{id}/return": {
            "put": {
                "tags": [
                    "Book Loans"
                ],
                "summary": "Mark book as returned",
                "description": "Mark a book loan as returned by setting returned_at timestamp",
                "operationId": "1f83fde52e5787a72c66985bbcf37e6c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book loan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book marked as returned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BookLoan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book loan not found"
                    }
                }
            }
        },
        "/api/loans/user/{userId}": {
            "get": {
                "tags": [
                    "Book Loans"
                ],
                "summary": "Get loans by user ID",
                "description": "Retrieve all book loans for a specific user",
                "operationId": "8b8ce308fe916f29daf57ee8f85da303",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User loans retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BookLoan"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "BookLoan": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "book_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "user_email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "book_title": {
                        "type": "string",
                        "example": "Laravel: Up & Running"
                    },
                    "book_author": {
                        "type": "string",
                        "example": "Matt Stauffer"
                    },
                    "loaned_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "returned_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "is_returned": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}